Fixed issue where a smaller ocean is being displayed when Ocean is disabled in menu.
This is honestly the only way I can figure out how to handle this situation. This is because "sealevel_mod" ends up being used in the shader code, and I don't know how the values are being replicated into that namespace, but it is. And just amplifying the "off" mode from "0.0" to "-1.0" makes the oceans go away completely. I think this is because the "sealevel_mod" is being multiplied with "sealevel" in shader code in order to figure out displayed portions of the ocean. Apparently "0.0" was still visible in some places. I'm guessing these are places where the displacement of the seafloor was below 0.