A fix for the world's stupidest bug
plate.mask for detaching crust was getting set to "1". It should have been set to "0", which would have indicated that the crust no longer exists. So in other words, detachment never occurred. I only discovered this after realizing accretion was the foremost cause for conservation leaks. This itself required me to discover a separate bug that caused accretion to be disabled. In other words, the effect of two bugs was to cancel each other out.