fixed issue causing seas to disappear when no forces are act on a plate
Plate motion is determined using a weighted average of all forces acting on a plate. When no forces exist, a divide by zero error occurs, and plate motion is set to an invalid value. This invalid value propagates throughout the simulation, causing getNearestIds() to return null values. These null values then get coerced into a 0 value when it comes time to merge plates into a single map. This causes the map to fill with the 0th index of the problematic plate. If the 0th index is land, the oceans will disappear. If the 0th index is sea, the land will disappear. In effect, it causes the world to take on a uniform appearance.

To resolve this issue, plate motion is set to the identity matrix whenever invalid values occur.