assuming atmo scale height is scalar for performance
We will assume that any planet whose atmosphere is thick enough to be require raymarching will probably have a turbopause that's high enough to encompass the entire atmosphere. Within the turbopause, atmospheric compounds will have intermixed to the point where they are all described by the same scale height. On Earth, the turbopause coincides with the official edge of space, so this appears to be a reasonable assumption