exposing vector representation of many<T>, removing many<T>.resize()
many<T>.resize() doesn't make sense since the intended use case for many<T> requires it to remain of constant size. We only use it in edge cases within the constructor for the Grid class. We remove it in favor of exposing the internal vector representation of many<T>. This is better since it also frees us up from having to completely reimplement the std::vector interface.